requirements.txt
tk
numpy
tensorflow
Pillow



